[WEJBHTTP-152] Refactoring - introducing ServerHandlers class
